Изменяет лейаут, добавляет index.jsx компонентам
