+++ src/common/components/Alert/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Alert'
