+++ src/common/components/Calendar/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Calendar'
