+++ src/common/components/Confirm/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Confirm'
