+++ src/common/components/ContentBox/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './ContentBox'
