+++ src/common/components/FacultySelector/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './FacultySelector'
