+++ src/common/components/FormModal/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './FormModal'
