+++ src/common/components/LinkContainer/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './LinkContainer'
