+++ src/common/components/Page/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Page'
