+++ src/modules/queue/components/QueueInputGroup/QueueInputGroup.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { Row, Col } from 'react-bootstrap'
-import MonthSelector from '@/common/components/MonthSelector/MonthSelector'
-import SlotCalendar from '@/modules/queue/components/SlotCalendar/SlotCalendar'
-import SliderCalendar from '@/modules/queue/components/SliderCalendar/SliderCalendar'
+import MonthSelector from '@/common/components/MonthSelector'
+import SlotCalendar from '@/modules/queue/components/SlotCalendar'
+import SliderCalendar from '@/modules/queue/components/SliderCalendar'
 import NamedSlotList from '@/modules/queue/components/SlotList/NamedSlotList'
 import { toObject, toDate, toDatetime } from '@/common/utils/dateLib'
 import useMediaQuery from '@/common/hooks/useMediaQuery'
