+++ src/modules/queue/components/QueueInputGroup/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './QueueInputGroup'
