+++ src/modules/queue/components/SliderCalendar/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SliderCalendar'
