+++ src/modules/queue/components/Slot/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Slot'
