+++ src/modules/queue/components/SlotCalendar/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SlotCalendar'
