+++ src/modules/queue/components/SlotList/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './SlotList'
