+++ src/pages/copyright.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Col, Row } from 'react-bootstrap'
-import Icon from '@/common/components/Icon/index'
-import Page from '@/common/components/Page/Page'
+import Icon from '@/common/components/Icon'
+import Page from '@/common/components/Page'
 
 export default function Copyright() {
     return (
